--- htdocs_symfony/src/Controller/App/CachesController.php
@@ -1,14 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Oc\Controller\App;
-
-/**
- * Class CachesController
- *
- * @package Oc\Controller\Backend
- */
-class CachesController extends \Oc\Controller\Backend\CachesController
-{
-}
