+++ htdocs_symfony/src/Controller/App/CachesControllerApp.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Oc\Controller\App;
+
+/**
+ * Class CachesControllerApp
+ *
+ */
+class CachesControllerApp extends \Oc\Controller\CachesControllerBase
+{
+}
